+++ fireball/src/pe/asm.rs
@@ -5,15 +5,11 @@ use capstone::Instructions;
 
 impl PE {
     /// 어셈블리 코드를 파싱한다.
-    pub(crate) fn parse_assem_range(
-        &self,
-        offset: Address,
-        size: usize,
-    ) -> Result<Instructions, ()> {
+    pub(crate) fn parse_assem_range(&self, offset: Address, size: u64) -> Result<Instructions, ()> {
         let file_offset = offset.get_file_offset();
         let virtual_offset = offset.get_virtual_address();
         let insns = match self.capstone.disasm_all(
-            &self.binary[file_offset..file_offset + size],
+            &self.binary[file_offset as usize..(file_offset + size) as usize],
             virtual_offset as u64,
         ) {
             Ok(insts) => insts,
