+++ fireball/src/pe/fire/decom_from_file_offset.rs
@@ -6,7 +6,7 @@ impl PE {
         &self,
         address: u64,
     ) -> Result<(), Box<dyn std::error::Error>> {
-        let _address = Address::from_file_offset(&self.binary, address as usize);
+        let _address = Address::from_file_offset(&self.binary, address);
         todo!();
     }
 }
