+++ fireball/src/pe/fire/decom_from_virtual_address.rs
@@ -6,7 +6,7 @@ impl PE {
         &self,
         address: u64,
     ) -> Result<(), Box<dyn std::error::Error>> {
-        let _address = Address::from_virtual_address(&self.binary, address as usize);
+        let _address = Address::from_virtual_address(&self.binary, address);
         todo!();
     }
 }
